requirements.txt
Flask==2.3.3
pypdf==3.17.0
Pillow==10.0.1
img2pdf==0.4.4
Werkzeug==2.3.7
gunicorn==21.2.0
